// extensions/spotify-music-player/src/helpers/spotify.api.ts
export type ExternalUrlObject = {
  spotify?: string;
};
export type ImageObject = {
  url: string;
  height: number | null;
  width: number | null;
};
export type AlbumRestrictionObject = {
  reason?: "market" | "product" | "explicit";
};
export type AlbumBase = {
  album_type: "album" | "single" | "compilation";
  total_tracks: number;
  available_markets: string[];
  external_urls: ExternalUrlObject;
  href: string;
  id: string;
  images: ImageObject[];
  name: string;
  release_date: string;
  release_date_precision: "year" | "month" | "day";
  restrictions?: AlbumRestrictionObject;
  type: "album";
  uri: string;
};
export type SimplifiedArtistObject = {
  external_urls?: ExternalUrlObject;
  href?: string;
  id?: string;
  name?: string;
  type?: "artist";
  uri?: string;
};
export type PagingObject = {
  href: string;
  limit: number;
  next: string | null;
  offset: number;
  previous: string | null;
  total: number;
};
export type SimplifiedAlbumObject = AlbumBase & {
  album_group?: "album" | "single" | "compilation" | "appears_on";
  artists: SimplifiedArtistObject[];
};
export type LinkedTrackObject = {
  external_urls?: ExternalUrlObject;
  href?: string;
  id?: string;
  type?: string;
  uri?: string;
};
export type TrackRestrictionObject = {
  reason?: string;
};
export type SimplifiedTrackObject = {
  album?: SimplifiedAlbumObject;
  artists?: SimplifiedArtistObject[];
  available_markets?: string[];
  disc_number?: number;
  duration_ms?: number;
  explicit?: boolean;
  external_urls?: ExternalUrlObject;
  href?: string;
  id?: string;
  is_playable?: boolean;
  linked_from?: LinkedTrackObject;
  restrictions?: TrackRestrictionObject;
  name?: string;
  preview_url?: string;
  track_number?: number;
  type?: string;
  uri?: string;
  is_local?: boolean;
};
export type PagingSimplifiedTrackObject = PagingObject & {
  items?: SimplifiedTrackObject[];
};
export type ExternalIdObject = {
  isrc?: string;
  ean?: string;
  upc?: string;
};
export type CopyrightObject = {
  text?: string;
  type?: string;
};
export type AlbumObject = AlbumBase & {
  artists?: SimplifiedArtistObject[];
  tracks?: PagingSimplifiedTrackObject;
  popularity?: number;
  label?: string;
  external_ids?: ExternalIdObject;
  genres?: string[];
  copyrights?: CopyrightObject[];
};
export type ErrorObject = {
  status: number;
  message: string;
};
export type FollowersObject = {
  href?: string | null;
  total?: number;
};
export type ArtistObject = {
  external_urls?: ExternalUrlObject;
  followers?: FollowersObject;
  genres?: string[];
  href?: string;
  id?: string;
  images?: ImageObject[];
  name?: string;
  popularity?: number;
  type?: "artist";
  uri?: string;
};
export type PagingSimplifiedAlbumObject = PagingObject & {
  items?: SimplifiedAlbumObject[];
};
export type TrackObject = {
  album?: SimplifiedAlbumObject;
  artists?: ArtistObject[];
  available_markets?: string[];
  disc_number?: number;
  duration_ms?: number;
  explicit?: boolean;
  external_ids?: ExternalIdObject;
  external_urls?: ExternalUrlObject;
  href?: string;
  id?: string;
  is_playable?: boolean;
  linked_from?: LinkedTrackObject;
  restrictions?: TrackRestrictionObject;
  name?: string;
  popularity?: number;
  preview_url?: string;
  track_number?: number;
  type?: "track";
  uri?: string;
  is_local?: boolean;
};